Refactor credential packages for incoming HCP support (#312)
