+++ internal/credentials/provider/provider.go
@@ -1,0 +1,17 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+package provider
+
+import (
+	"context"
+
+	"k8s.io/apimachinery/pkg/types"
+	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+type CredentialProviderBase interface {
+	GetUID() types.UID
+	GetNamespace() string
+	GetCreds(context.Context, ctrlclient.Client) (map[string]interface{}, error)
+}
