+++ internal/credentials/vault/aws_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package credentials
+package vault
 
 import (
 	"fmt"
