+++ internal/credentials/vault/consts.go
@@ -1,0 +1,13 @@
+// Copyright (c) HashiCorp, Inc.
+// SPDX-License-Identifier: MPL-2.0
+
+package vault
+
+const (
+	ProviderSecretKeyAppRole = "id"
+	ProviderSecretKeyJWT     = "jwt"
+	ProviderMethodKubernetes = "kubernetes"
+	ProviderMethodJWT        = "jwt"
+	ProviderMethodAppRole    = "appRole"
+	ProviderMethodAWS        = "aws"
+)
