+++ internal/helpers/hmac_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package vault
+package helpers
 
 import (
 	"crypto/rand"
