+++ internal/vault/cache_key_test.go
@@ -13,7 +13,7 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 
 	secretsv1beta1 "github.com/hashicorp/vault-secrets-operator/api/v1beta1"
-	"github.com/hashicorp/vault-secrets-operator/internal/vault/credentials"
+	"github.com/hashicorp/vault-secrets-operator/internal/credentials/vault"
 )
 
 const (
@@ -233,7 +233,7 @@ func TestComputeClientCacheKeyFromClient(t *testing.T) {
 				c = &defaultClient{
 					authObj: tt.authObj,
 					connObj: tt.connObj,
-					credentialProvider: credentials.NewKubernetesCredentialProvider(nil, "",
+					credentialProvider: vault.NewKubernetesCredentialProvider(nil, "",
 						tt.providerUID),
 				}
 			}
@@ -257,21 +257,21 @@ func TestClientCacheKey_IsClone(t *testing.T) {
 		{
 			name: "is-not-a-clone-no-suffix",
 			k: ClientCacheKey(fmt.Sprintf("%s-%s",
-				credentials.ProviderMethodKubernetes,
+				vault.ProviderMethodKubernetes,
 				computedHash)),
 			want: false,
 		},
 		{
 			name: "is-not-a-clone-empty-suffix",
 			k: ClientCacheKey(fmt.Sprintf("%s-%s-",
-				credentials.ProviderMethodKubernetes,
+				vault.ProviderMethodKubernetes,
 				computedHash)),
 			want: false,
 		},
 		{
 			name: "is-a-clone",
 			k: ClientCacheKey(fmt.Sprintf("%s-%s-ns1/ns2",
-				credentials.ProviderMethodKubernetes,
+				vault.ProviderMethodKubernetes,
 				computedHash)),
 			want: true,
 		},
@@ -295,18 +295,18 @@ func TestClientCacheKeyClone(t *testing.T) {
 		{
 			name: "valid",
 			key: ClientCacheKey(fmt.Sprintf("%s-%s",
-				credentials.ProviderMethodKubernetes,
+				vault.ProviderMethodKubernetes,
 				computedHash)),
 			namespace: "ns1/ns2",
 			want: ClientCacheKey(fmt.Sprintf("%s-%s-ns1/ns2",
-				credentials.ProviderMethodKubernetes,
+				vault.ProviderMethodKubernetes,
 				computedHash)),
 			wantErr: assert.NoError,
 		},
 		{
 			name: "fail-empty-namespace",
 			key: ClientCacheKey(fmt.Sprintf("%s-%s",
-				credentials.ProviderMethodKubernetes,
+				vault.ProviderMethodKubernetes,
 				computedHash)),
 			namespace: "",
 			want:      "",
@@ -317,7 +317,7 @@ func TestClientCacheKeyClone(t *testing.T) {
 		{
 			name: "fail-parent-is-clone",
 			key: ClientCacheKey(fmt.Sprintf("%s-%s-ns1/ns2",
-				credentials.ProviderMethodKubernetes,
+				vault.ProviderMethodKubernetes,
 				computedHash)),
 			namespace: "ns3",
 			want:      "",
