+++ internal/vault/cache_storage.go
@@ -28,6 +28,7 @@ import (
 	secretsv1beta1 "github.com/hashicorp/vault-secrets-operator/api/v1beta1"
 	"github.com/hashicorp/vault-secrets-operator/internal/common"
 	"github.com/hashicorp/vault-secrets-operator/internal/consts"
+	"github.com/hashicorp/vault-secrets-operator/internal/helpers"
 	"github.com/hashicorp/vault-secrets-operator/internal/metrics"
 )
 
@@ -248,7 +249,7 @@ func (c *defaultClientCacheStorage) Store(ctx context.Context, client ctrlclient
 	}
 
 	var messageMAC []byte
-	messageMAC, err = macMessage(c.hmacKey, message)
+	messageMAC, err = helpers.MACMessage(c.hmacKey, message)
 	if err != nil {
 		return nil, err
 	}
@@ -513,7 +514,7 @@ func (c *defaultClientCacheStorage) validateSecretMAC(req ClientCacheStorageRest
 		return err
 	}
 
-	ok, _, err = validateMAC(message, messageMAC, c.hmacKey)
+	ok, _, err = helpers.ValidateMAC(message, messageMAC, c.hmacKey)
 	if err != nil {
 		return err
 	}
@@ -610,7 +611,7 @@ func NewDefaultClientCacheStorage(ctx context.Context, client ctrlclient.Client,
 		return nil, err
 	}
 
-	s, err := createHMACKeySecret(ctx, client, config.HMACSecretObjKey)
+	s, err := helpers.CreateHMACKeySecret(ctx, client, config.HMACSecretObjKey)
 	if err != nil {
 		if !apierrors.IsAlreadyExists(err) {
 			return nil, err
@@ -618,7 +619,7 @@ func NewDefaultClientCacheStorage(ctx context.Context, client ctrlclient.Client,
 	}
 
 	if s == nil {
-		s, err = getHMACKeySecret(ctx, client, config.HMACSecretObjKey)
+		s, err = helpers.GetHMACKeySecret(ctx, client, config.HMACSecretObjKey)
 		if err != nil {
 			return nil, err
 		}
@@ -626,7 +627,7 @@ func NewDefaultClientCacheStorage(ctx context.Context, client ctrlclient.Client,
 
 	cacheStorage := &defaultClientCacheStorage{
 		hmacSecretObjKey:  config.HMACSecretObjKey,
-		hmacKey:           s.Data[hmacKeyName],
+		hmacKey:           s.Data[helpers.HMACKeyName],
 		enforceEncryption: config.EnforceEncryption,
 		logger:            zap.New().WithName("ClientCacheStorage"),
 		requestCounterVec: prometheus.NewCounterVec(
