+++ internal/vault/client.go
@@ -17,8 +17,9 @@ import (
 	secretsv1beta1 "github.com/hashicorp/vault-secrets-operator/api/v1beta1"
 	"github.com/hashicorp/vault-secrets-operator/internal/common"
 	"github.com/hashicorp/vault-secrets-operator/internal/consts"
+	"github.com/hashicorp/vault-secrets-operator/internal/credentials"
+	"github.com/hashicorp/vault-secrets-operator/internal/credentials/provider"
 	"github.com/hashicorp/vault-secrets-operator/internal/metrics"
-	"github.com/hashicorp/vault-secrets-operator/internal/vault/credentials"
 )
 
 type ClientOptions struct {
@@ -147,7 +148,7 @@ type Client interface {
 	Validate() error
 	GetVaultAuthObj() *secretsv1beta1.VaultAuth
 	GetVaultConnectionObj() *secretsv1beta1.VaultConnection
-	GetCredentialProvider() credentials.CredentialProvider
+	GetCredentialProvider() provider.CredentialProviderBase
 	GetCacheKey() (ClientCacheKey, error)
 	Close()
 	Clone(string) (Client, error)
@@ -167,7 +168,7 @@ type defaultClient struct {
 	skipRenewal        bool
 	lastRenewal        int64
 	targetNamespace    string
-	credentialProvider credentials.CredentialProvider
+	credentialProvider provider.CredentialProviderBase
 	watcher            *api.LifetimeWatcher
 	lastWatcherErr     error
 	once               sync.Once
@@ -248,7 +249,7 @@ func (c *defaultClient) Clone(namespace string) (Client, error) {
 	return client, nil
 }
 
-func (c *defaultClient) GetCredentialProvider() credentials.CredentialProvider {
+func (c *defaultClient) GetCredentialProvider() provider.CredentialProviderBase {
 	return c.credentialProvider
 }
