--- internal/vault/credentials/consts.go
@@ -1,10 +1,0 @@
-// Copyright (c) HashiCorp, Inc.
-// SPDX-License-Identifier: MPL-2.0
-
-package credentials
-
-const (
-	TokenGenerateName        = "vso-"
-	ProviderSecretKeyAppRole = "id"
-	ProviderSecretKeyJWT     = "jwt"
-)
